Fixed build after sync with master
