--- src/main/java/com/chargify/model/IntervalUnit.java
@@ -1,6 +1,0 @@
-package com.chargify.model;
-
-public enum IntervalUnit
-{
-  month, day
-}
